requirements.txt
python-telegram-bot==20.7
sqlalchemy==1.4.52
python-dotenv==1.0.1
alembic==1.13.1
